Removes nearly all dynamic allocations.
Use of std::vector has been almost entirely replaced by std::array.
In every case there is a small upper bound to the max array size
for e.g. a deck of cards, tableau, foundation, move extras. These
can all be allocated on the stack with no need for allocating and
freeing at a very high rate like it was before. String streams have
also been eradicated for the most part since valgrind showed that was
a major slowdown.

The number of moves per second is now around 150000 whereas before this
optimization pass it was around 9000 (compiling with -Ofast in both
cases). The javascript version gets around 4000 moves/sec. So this is
an order of magnitude speedup and then some.